DDR: Loop to find proper super type for DIE
When dealing with (but perhaps not limited to) the split dwarf debug
files (.dwo) produced with Open XL on z/OS, there are instances where
the inheritance type points to a typedef Die. This updates the
getSuperUDT() logic to keep iterating until it finds the real base/super
type the current Die is supposed to inherit from.

Signed-off-by: Gaurav Chaudhari <Gaurav.Chaudhari@ibm.com>